# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/books": {
      "get": {
        "description": "支持 ?page&pageSize 分页（data 为 { list,total,page,pageSize }）；默认返回数组。支持 status/category/keyword 过滤。",
        "operationId": "BooksController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码，从 1 开始",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "按阅读状态过滤",
            "schema": {
              "type": "string",
              "enum": [
                "reading",
                "completed",
                "want-to-read"
              ]
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "按分类过滤",
            "schema": {
              "example": "计算机",
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "关键词模糊匹配 title/author/tags",
            "schema": {
              "example": "CSS",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "统一信封 { code:0, message:\"ok\", data }；data 为 BookListResultDto（分页时）或 BookDto[]（未分页时）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookListResultDto"
                }
              }
            }
          }
        },
        "summary": "书籍列表",
        "tags": [
          "books"
        ]
      }
    },
    "/api/books/{id}": {
      "get": {
        "description": "返回单个 Book（含 bookPages）",
        "operationId": "BooksController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "书籍 ID",
            "schema": {
              "example": "design-psychology",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "统一信封 { code:0, message:\"ok\", data }；data 为 BookDto",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDto"
                }
              }
            }
          }
        },
        "summary": "书籍详情",
        "tags": [
          "books"
        ]
      }
    },
    "/api/books/{id}/progress": {
      "post": {
        "description": "前端翻页/离开阅读器时上报，后端 upsert 到 SQLite；列表/详情接口返回时用 DB 实时值覆盖文件 frontmatter 的初始 progress。",
        "operationId": "BooksController_updateProgress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "书籍 ID",
            "schema": {
              "example": "java-in-action-2e",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookProgressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "统一信封 { code:0, message:\"ok\", data }；data 为保存后的 BookProgress",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookProgress"
                }
              }
            }
          }
        },
        "summary": "上报阅读进度",
        "tags": [
          "books"
        ]
      }
    },
...
  },
  "info": {
    "title": "个人博客 API",
    "description": "覆盖学习文章(posts) / 兴趣视频(videos) / 书籍(books) / 项目 / 时间线 / 评论 / 统计，统一响应 { code, message, data }",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "schemas": {
...
      "BookPageItemDto": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "number",
            "description": "页码，从 1 开始"
          },
          "title": {
            "type": "string",
            "description": "页面标题"
          },
          "chapter": {
            "type": "string",
            "description": "章节名"
          },
          "type": {
            "type": "string",
            "description": "页面类型",
            "enum": [
              "cover",
              "copyright",
              "content",
              "illustration",
              "notes",
              "back-cover",
              "pdf-page",
              "code-page",
              "epub-section"
            ]
          },
          "content": {
            "type": "string",
            "description": "正文（Markdown 字符串）"
          },
          "format": {
            "type": "string",
            "description": "格式",
            "example": "markdown"
          },
          "image": {
            "type": "string",
            "description": "图片 URL"
          },
          "pdfPageNumber": {
            "type": "number",
            "description": "PDF 影印原页号"
          },
          "pdfUrl": {
            "type": "string",
            "description": "PDF 页源文件"
          },
          "codeLanguage": {
            "type": "string",
            "description": "代码页语言"
          },
          "codeSnippet": {
            "type": "string",
            "description": "代码内容"
          },
          "quote": {
            "type": "string",
            "description": "引言/金句"
          }
        },
        "required": [
          "pageNumber",
          "type"
        ]
      },
      "BookTocItemDto": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "number",
            "description": "该章节在 bookPages 中的起始页码"
          },
          "title": {
            "type": "string",
            "description": "章节/部分标题",
            "example": "第 1 章　Java 8、9、10以及11的变化"
          },
          "level": {
            "type": "number",
            "description": "层级：1=部分/章，2=节，3=小节，4=四级标题",
            "example": 2
          }
        },
        "required": [
          "pageNumber",
          "title",
          "level"
        ]
      },
      "BookDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "唯一 ID",
            "example": "design-psychology"
          },
          "title": {
            "type": "string",
            "description": "书名"
          },
          "author": {
            "type": "string",
            "description": "作者"
          },
          "coverUrl": {
            "type": "string",
            "description": "封面图 URL"
          },
          "category": {
            "type": "string",
            "description": "分类",
            "example": "计算机"
          },
          "rating": {
            "type": "number",
            "description": "评分（0–5 整数）"
          },
          "status": {
            "type": "string",
            "description": "阅读状态",
            "enum": [
              "reading",
              "completed",
              "want-to-read"
            ]
          },
          "progress": {
            "type": "number",
            "description": "阅读进度 0–100"
          },
          "lastPage": {
            "type": "number",
            "description": "最近阅读的 bookPages 页码（断点续读，来自 DB 实时进度）",
            "example": 38
          },
          "summary": {
            "type": "string",
            "description": "简介"
          },
          "excerpt": {
            "type": "string",
            "description": "原文章节节选（Markdown）"
          },
          "thoughts": {
            "type": "string",
            "description": "深度读书笔记（Markdown）"
          },
          "tags": {
            "description": "标签",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "publishYear": {
            "type": "string",
            "description": "出版年份",
            "example": "2024"
          },
          "pages": {
            "type": "number",
            "description": "纸质页数"
          },
          "spineColor": {
            "type": "string",
            "description": "3D 书脊渐变色（tailwind 类）",
            "example": "from-amber-600 to-orange-700"
          },
          "formats": {
            "description": "支持格式",
            "example": [
              "markdown",
              "pdf"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "pdfUrl": {
            "type": "string",
            "description": "真实 PDF 文件链接"
          },
          "epubUrl": {
            "type": "string",
            "description": "真实 EPUB 文件链接"
          },
          "bookPages": {
            "description": "阅读器渲染页（不传则前端自动生成 8 页默认结构）",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookPageItemDto"
            }
          },
          "toc": {
            "description": "目录（章节 → 起始页码），可用于前端导航",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookTocItemDto"
            }
          }
        },
        "required": [
          "id",
          "title",
          "author",
          "coverUrl",
          "category",
          "rating",
          "status",
          "progress",
          "summary",
          "tags"
        ]
      },
      "BookListResultDto": {
        "type": "object",
        "properties": {
          "list": {
            "description": "当前页数据",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookDto"
            }
          },
          "total": {
            "type": "number",
            "description": "总条数"
          },
          "page": {
            "type": "number",
            "description": "当前页码"
          },
          "pageSize": {
            "type": "number",
            "description": "每页数量"
          }
        },
        "required": [
          "list",
          "total",
          "page",
          "pageSize"
        ]
      },
      "UpdateBookProgressDto": {
        "type": "object",
        "properties": {
          "progress": {
            "type": "number",
            "description": "阅读进度 0–100",
            "example": 45,
            "minimum": 0,
            "maximum": 100
          },
          "lastPage": {
            "type": "number",
            "description": "最近阅读的 bookPages 页码（断点续读用）",
            "example": 38,
            "minimum": 1
          }
        },
        "required": [
          "progress"
        ]
      },
      "BookProgress": {
        "type": "object",
        "properties": {
          "bookId": {
            "type": "string",
            "description": "书籍 ID",
            "example": "java-in-action-2e"
          },
          "progress": {
            "type": "number",
            "description": "阅读进度 0–100",
            "example": 45
          },
          "lastPage": {
            "type": "object",
            "description": "最近阅读的 bookPages 页码",
            "example": 38
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "最近上报时间"
          }
        },
        "required": [
          "bookId",
          "progress",
          "lastPage",
          "updatedAt"
        ]
      },
...
    }
  }
}