# backend/src/content/books/update-book-progress.dto.ts
import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
import { Type } from 'class-transformer';
import { IsInt, IsOptional, Max, Min } from 'class-validator';

export class UpdateBookProgressDto {
  @ApiProperty({ description: '阅读进度 0–100', example: 45, minimum: 0, maximum: 100 })
  @Type(() => Number)
  @IsInt()
  @Min(0)
  @Max(100)
  progress: number;

  @ApiPropertyOptional({
    description: '最近阅读的 bookPages 页码（断点续读用）',
    example: 38,
    minimum: 1,
  })
  @IsOptional()
  @Type(() => Number)
  @IsInt()
  @Min(1)
  lastPage?: number;
}